==> 3. Generate_Similarity.ipynb <==
from transformers import BertForSequenceClassification, BertTokenizer
import pandas as pd
import torch
from sklearn.metrics.pairwise import cosine_similarity, paired_distances
from scipy import spatial
from sentence_transformers import SentenceTransformer
from transformers import AutoTokenizer, AutoModel
import torch
import json

# !pip install torch
# !pip install transformers
# !pip install sentence_transformers

# Load model
tokenizer = AutoTokenizer.from_pretrained('anferico/bert-for-patents')
model = AutoModel.from_pretrained('anferico/bert-for-patents')

# this function splits the data
def get_split(text,length):
    l_total = []
    l_parcial = []
    n = len(text.split())//length
    i = 0
    while i < n:
        l_parcial = text.split()[i*length:(i+1)*length]
        l_total.append(" ".join(l_parcial))
        i+=1
        
    l_parcial = text.split()[i*length:]
    l_total.append(" ".join(l_parcial))
        
    return l_total


# this function is for pooling
def cls_pooling(model_output, attention_mask):
    return model_output[0][:,0]

# this function generates similarity score from two text
def get_sim_score(text1,text2,tokenizer,model):
    # prepare the input texts
    len_of_split = 300
    text1 = get_split(str(text1),len_of_split)
    text2 = get_split(str(text2),len_of_split)
    sep = len(text1)
    combined_text = text1+text2
    
    # generate the vector sets
    encoded_input = tokenizer(combined_text, max_length=512, padding=True, truncation=True, return_tensors='pt')
    with torch.no_grad():
        model_output = model(**encoded_input)
    combined_vec = cls_pooling(model_output, encoded_input['attention_mask'])
    
    # compare each part of two texts and calculate the similarity score
    p1 = combined_vec[:sep]
    p2 = combined_vec[sep:]
    similarity_score = 0
    for i in p1:
        for j in p2:
            cosine_sim = 1 - spatial.distance.cosine(i, j)
            if cosine_sim>similarity_score:
                similarity_score = cosine_sim


    return similarity_score


# this function transform the text in the patent dictionary into vec
def generate_vec_for_one(patent,tokenizer,model):
    # save all the splited texts
    content_all = []
    # record the key and length of splited texts
    ref_list = []
    # traverse the plaintiff
    patent_length = 0
    for key,value in patent.items():
        content = get_split(str(value),300)
        length = len(content)
        content_all+=content
        ref_list.append((key,length))
        patent_length+=length
    
    # calculate the vec
    encoded_input = tokenizer(content_all, max_length=512, padding=True, truncation=True, return_tensors='pt')
    with torch.no_grad():
        model_output = model(**encoded_input)
    content_vec = cls_pooling(model_output, encoded_input['attention_mask'])
    
    # write the vec back to dict
    patent_vec_dict = {}

    idx = 0
    i = 0
    while idx < patent_length: 
        key = ref_list[i][0]
        length = ref_list[i][1]
        content = content_vec[idx:idx+length]
        patent_vec_dict[key] = content
        idx +=length
        i+=1

    
    return patent_vec_dict


# this function generates a dataframe of similarity scores of different combination of each parts of plantiff and defendant patent 
def compare_two_patents(plaintiff_no,defendant_no,data_dict,tokenizer,model):
    score_dict = {'Plaintiff_Patent':[],'Defendant_Patent':[],'Plaintiff_Category':[],'Defendant_Category':[],'Similarity_Score':[]}
    plaintiff = data_dict[plaintiff_no]
    defendant = data_dict[defendant_no]
    p_vec, d_vec = generate_vec(plaintiff,defendant,tokenizer,model)
    
    # generate the df
    for keyp,valuep in p_vec.items():
        for keyd,valued in d_vec.items():

            # calculate the max for the combination
            sim_score = 0
            for vec1 in valuep:
                for vec2 in valued:
                    cosine_sim = 1 - spatial.distance.cosine(vec1, vec2)
                    if cosine_sim > sim_score:
                        sim_score = cosine_sim

            score_dict['Plaintiff_Patent'].append(plaintiff_no)
            score_dict['Defendant_Patent'].append(defendant_no)
            score_dict['Plaintiff_Category'].append(keyp)
            score_dict['Defendant_Category'].append(keyd)
            score_dict['Similarity_Score'].append(sim_score)
            
    score_df = pd.DataFrame(score_dict)
    return score_df

# this function use less memory but much longer time 
def compare_two_patents_less_memo(plaintiff_no,defendant_no,data_dict):
    score_dict = {'Plaintiff_Patent':[],'Defendant_Patent':[],'Plaintiff_Category':[],'Defendant_Category':[],'Similarity_Score':[]}
    plaintiff = data_dict[plaintiff_no]
    defendant = data_dict[defendant_no]
    for keyp,valuep in plaintiff.items():
        for keyd,valued in defendant.items():
            score_dict['Plaintiff_Patent'].append(plaintiff_no)
            score_dict['Defendant_Patent'].append(defendant_no)
            score_dict['Plaintiff_Category'].append(keyp)
            
            score_dict['Defendant_Category'].append(keyd)
            score = get_sim_score(valuep,valued,tokenizer,model)
            score_dict['Similarity_Score'].append(score)
            
    score_df = pd.DataFrame(score_dict)
    return score_df

# this function is the advanced version of the previous one with dict
def compare_two_patents_with_dict(plaintiff_no,defendant_no,data_dict,tokenizer,model):
    score_dict = {'Plaintiff_Patent':[],'Defendant_Patent':[],'Plaintiff_Category':[],'Defendant_Category':[],'Similarity_Score':[]}
    plaintiff = data_dict[plaintiff_no]
    defendant = data_dict[defendant_no]
    
    # generate the vec for patents
    if plaintiff_no in patent_vec_dict:
        p_vec = patent_vec_dict[plaintiff_no]
    else:
        p_vec = generate_vec_for_one(plaintiff,tokenizer,model)
        patent_vec_dict[plaintiff_no] = p_vec
        
    if defendant_no in patent_vec_dict:
        d_vec = patent_vec_dict[defendant_no]
    else:
        d_vec = generate_vec_for_one(defendant,tokenizer,model)
        patent_vec_dict[defendant_no] = d_vec
    
    
    # generate the df
    for keyp,valuep in p_vec.items():
        for keyd,valued in d_vec.items():

            # calculate the max for the combination
            sim_score = 0
            for vec1 in valuep:
                for vec2 in valued:
                    cosine_sim = 1 - spatial.distance.cosine(vec1, vec2)
                    if cosine_sim > sim_score:
                        sim_score = cosine_sim

            score_dict['Plaintiff_Patent'].append(plaintiff_no)
            score_dict['Defendant_Patent'].append(defendant_no)
            score_dict['Plaintiff_Category'].append(keyp)
            score_dict['Defendant_Category'].append(keyd)
            score_dict['Similarity_Score'].append(sim_score)
            
    score_df = pd.DataFrame(score_dict)
    return score_df

# load data dictionary
with open("patent_data.json",'r', encoding='UTF-8') as f:
     patent_data = json.load(f)

# load the plantiff and defendant patent no.
df_train_no = pd.read_csv('Simplified_training_data_new.csv')

# generate the df for the first 50 pairs
df_train_no = df_train_no[350:370]
df_train_no

# set the dict to score vec
patent_vec_dict = {}

# the final output dataframe
score_dict = {'Plaintiff_Patent':[],'Defendant_Patent':[],'Plaintiff_Category':[],'Defendant_Category':[],'Similarity_Score':[]}
score_df = pd.DataFrame(score_dict)
# calculate the similarity dataframe for each pair and append them to the score_df
n = 1
for plaintiff_no, defendant_no in zip(df_train_no['plaintiff_patent'],df_train_no['defendant_patent']):
    print('Now start the ',n,' th pair calculation.')
    new_score_df = compare_two_patents_with_dict(plaintiff_no,defendant_no,patent_data,tokenizer,model)
    score_df = score_df.append(new_score_df, ignore_index=True)
    print(n,'th has finished.')
    n+=1


score_df

# save the score dataframe to csv
score_df.to_csv('Pairs_Part_Scores350-369.csv')

len(patent_data['US8216209B2']['Figures_Desc'])

# this part use to clean the similarity vectors for the patent with float 'nan' Abstract
import numpy as np

def get_useless_set(patent_data):
    useless_set = set()
    for key in patent_data:
        if type(patent_data[key]['Abstract'])==float:
            useless_set.add(key)
    return useless_set

def clean_abstract(df,useless_set):
    mask1 = []
    mask2 = []
    for i in range(len(df)):
        if (df.iloc[i]['Plaintiff_Patent'] in useless_set) and (df.iloc[i]['Plaintiff_Category']=='Abstract'):
            mask1.append(True)
        else:
            mask1.append(False)

        if (df.iloc[i]['Defendant_Patent'] in useless_set) and (df.iloc[i]['Defendant_Category']=='Abstract'):
            mask2.append(True)
        else:
            mask2.append(False)
        
    mask = np.array(mask1)+np.array(mask2)
    
    df_new = df[~mask]
    
    return df_new
    

df_needtoclean = pd.read_csv('Pairs_Part_Scores0-999.csv')
useless_set = get_useless_set(patent_data)
df_clean = clean_abstract(df_needtoclean,useless_set)

df_clean

df_clean.to_csv('test_train_data.csv')
